ECP-1119: First attempt, using ORM hints to speed up.
So, we have 'download_incoming_invoices' as a fetch group, that we enable using
fetchPlan.addGroup(...) in order to activate and ensure that the matching
incoming invoice items also have their reversalOf and charges eagerly loaded.

To do this requires changing JDOQL queries into typesafe queries in
IncomingInvoiceItemRepository.

We also realised we could remove a query against IncomingInvoiceRepository
(to find not-NEW invoices that have at least one item that hasn't been
reported), and just push up that not-NEW predicate into the original query
to find incoming invoice items.

However, we are still left with an N+1 issue, because as we process the
incoming invoice items with their respective invoices, we see an N+1 when
resolving invoice.property.  Trying to set a 'download_incoming_invoices' as
a fetch group for IncomingInvoice does not work.

We've therefore decided that it is going to be more robust to simply use a
view model, in the style of InvoiceSummaryForInvoiceRun"